+++ src/components/content/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./content.component";
+export type { ContentProps } from "./content.component";